+++ app/src/main/java/com/vivy/test/searchmydoctor/network/ApiNetworkManager.kt
@@ -1,5 +1,6 @@
 package com.vivy.test.searchmydoctor.network
 
+import com.vivy.test.searchmydoctor.BuildConfig
 import com.vivy.test.searchmydoctor.event.LoginTokenEvent
 import com.vivy.test.searchmydoctor.event.RequestFailureEvent
 import com.vivy.test.searchmydoctor.event.SearchSuccessEvent
@@ -8,6 +9,7 @@ import io.reactivex.android.schedulers.AndroidSchedulers
 import io.reactivex.disposables.Disposable
 import io.reactivex.schedulers.Schedulers
 import okhttp3.OkHttpClient
+import okhttp3.logging.HttpLoggingInterceptor
 import retrofit2.Retrofit
 import retrofit2.adapter.rxjava2.RxJava2CallAdapterFactory
 import retrofit2.converter.gson.GsonConverterFactory
@@ -54,7 +56,7 @@ class ApiNetworkManager(private val loginClient : OkHttpClient, private val sear
                 )
     }
 
-    override fun searchDoctorByName(docName: String, latitude: Float, longitude: Float): Disposable {
+    override fun searchDoctorByName(docName: String, latitude: Double, longitude: Double): Disposable {
         return searchApi
                 .searchDoctorByName(docName, latitude, longitude)
                 .subscribeOn(Schedulers.io())
@@ -65,7 +67,7 @@ class ApiNetworkManager(private val loginClient : OkHttpClient, private val sear
                 )
     }
 
-    override fun searchAllDoctors(latitude: Float, longitude: Float): Disposable {
+    override fun searchAllDoctors(latitude: Double, longitude: Double): Disposable {
         return searchApi
                 .searchAllDoctors(latitude, longitude)
                 .subscribeOn(Schedulers.io())
