+++ app/src/main/java/com/vivy/test/searchmydoctor/network/NetworkManager.kt
@@ -7,6 +7,6 @@ import retrofit2.Callback
 
 interface NetworkManager {
     fun login(username: String, password: String): Disposable
-    fun searchDoctorByName(docName: String, latitude: Float, longitude: Float): Disposable
-    fun searchAllDoctors(latitude: Float, longitude: Float): Disposable
+    fun searchDoctorByName(docName: String, latitude: Double, longitude: Double): Disposable
+    fun searchAllDoctors(latitude: Double, longitude: Double): Disposable
 }
